Document the cekit version we are using
Cekit is a fast-moving project. If you build the same image sources with two different versions of cekit you might get very different outputs when compared with e.g. diff or git diff. It's therefore helpful for us to document which version we are using at the present moment so we can manage the diff caused by tooling changes when we review newer versions of the tooling.