gendocs: generate image ENV var documentation
gendocs.py is a rudimentary Python script that operates on the output of
a cekit build (including a dry-run) and generates AsciiDoc
Environment-variable documentation for the image.

gendocs.sh is a harness to invoke gendocs.py for every image and version
we care about: it switches to different git branches and tags, collects
the image descriptors and runs gendocs.py.

Add a GitHub workflow to invoke gendocs.sh and upload the results to
GitHub Pages. Trigger on pushes to the ubi8 or ubi9 branches, or a tag
matching the pattern for releases.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>