Remove private repo URI from redhat overrides and GitHub workflows
Our internal build infrastructure now supports
registry.access.redhat.com in the FROM line. We no longer need to
override that for internal builds.

We don't need to pull-and-tag the public URI in the GitHub workflows (in
fact we haven't needed to since the FROM line was changed to the public
URI in the main image descriptors)