[OPENJDK-312] Don't chmod /etc/passwd
CRI-O on OCP >= 4.2 generates a passwd line at runtime for the
running user's random UID, which removes the requirement for us
to do the same.

Fixes CVE-2021-20264.

Use nss_wrapper to provide passwd metadata for the running UID,
based on the template passwd line for the jboss user. Necessary
for OCP 3.11.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>