Tests: de-couple Jolokia testing from other Behave features
Ports check was erroneously checking that the Jolokia port was open
(and wasn't checking 8443). Fix that and remove the Jolokia metadata
port check (which is already duplicated over in the Jolokia feature)

Don't check for Jolokia in container log test in the runtime feature.

Adjust the Jolokia test tagging to cover the correct OpenJDK images.
Also: fix the test! IT did not specify the starting conditions for a
container.

Add a new Jolokia feature that replicates the "runtime" test but
only the Jolokia-specific log check.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>