[OPENJDK-493] Don't install NSS Wrapper for OpenJDK17 containers
Introduce a new module jboss.container.util.nss-wrapper to encapsulate
the nss_wrapper config introduced with OPENJDK-312.

Move (almost) all nss_wrapper configuration into the new module.
(The configure_passwd function in run-java.sh remains, however, it's
benign. To remove this we need to extend run-java to support pluggable
functionality.)

Add the new module as an explicit dependency to the images which use
it, that is, all of the full-fat images.

No changes made to the OpenJDK 17 descriptor, which means it should
omit the nss_wrapper configuration entirely.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>