Rename jboss.container.java.run{,.bash} module
Remove the ".bash" suffix from the jboss.container.java.run module.

The ".bash" suffixes date from a time when we were open to the
possibility of implementing some container configuration in languages
other than shell, e.g., Python. However, with a focus on container
image sizes, this is no longer going to happen.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>