OPENJDK-533: test $HOME/passwd can be written to
run-java.sh has a line which attempts to write to $HOME/passwd, which is
part of the solution for backwards compatibility on OCP 3.11
(OPENJDK-312). This was made writeable by a common module but that was
moved to a new nss-wrapper module in OPENJDK-493, which is not installed
in OpenJDK 17 containers.

Two problems result for the OpenJDK-17 image:

 • Starting a downstream (S2I) container with a random UID (OpenShift
   default) terminates when the attempt to write to /home/jboss/passwd
   fails
 • for some container runtimes (at least Docker), the container will
   refuse to start if the running uid can't chdir to /home/jboss

That change also moved changing the /home/jboss permissions. Partially
revert those changes and check to ensure the file can be written before
attempting to.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>