Remove dependency between s2i.bash module and prometheus
This is not a hard dependency. The use of the variables declared by the
prometheus module in the s2i run script is guarded by a test that will
correctly fail if the prometheus module is not installed.

This has no impact on the selection or ordering of the modules for any
of the images.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>