[OPENJDK-559] Set MaxRAMPercentage=80 by default
Re-write the handling of JAVA_MAX_MEM_RATIO to set -XX:MaxRAMPercentage,
rather than try to set -Xmx and do calculations based on cgroups memory
limits (leave that up to the JVM)

Re-write JAVA_INITIAL_MEM_RATIO similarly. Do not specify a default
value for JAVA_INITIAL_MEM_RATIO: only handle it if the user sets it.
Since this is just a thin wrapper around setting
-XX:InitialRAMPercentage directly, the user can achieve this directly
with JAVA_OPTS. Mark this as deprecated, to be removed in a future
release.

Re-write the handling of JAVA_MAX_MEM_RATIO. Do not set a default
value. If the user specifies a value, map it directly to -Xms. As
with JAVA_INITIAL_MEM_RATIO, this can be achieved directly using
JAVA_OPTS, so mark this as deprecated for removal in a future
release.

If JAVA_MAX_INITIAL_MEM (-Xms) is set, then -XX:InitialRAMPercentage
would be ignored, so do not set it.

We are now no longer using any value derived from interpreting
cgroups v1 files, so remove the container-limits script entirely
and all references to CONTAINER_MAX_MEMORY or CONTAINER_CORE_LIMIT.

Add tests for the default and configured values.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>