Rename maven module for OpenJDK 11 to be explicit
The "maven" module version 8.2.3.6 installed the "maven" RPM, pulling
in whatever the default system JDK is implicitly.

Align the module with the other two, by re-versioning it to 8.6.3.6.11
(+ .11 suffix for JDK 11, change 8.2 → 8.6 to reflect the RHEL minor
version where we are implementing this) and depending explicitly on
maven-openjdk11.

depend on this module in ubi8-openjdk-11.yaml, and remove singleton-jdk
which should not be necessary.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>