Fix and enhance Behave GC tests
The Behave GC tests were broken by changes to (unrelated)
JAVA_MAX_MEM_RATIO. The tests are too sensitive because they check
the context around the arguments they are concerned about and are
sensitive to ordering.

Re-work the tests to only check for the parameter they are testing.
Whilst we're at it, re-work them to not require S2I, either, and
re-tag them to run on the runtime images too, now that they can.

This should also reduce the time required to check them.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>