Update Maven modules to 3.8, simplify version scheme
Update all references to the 3.6 AppStream for Maven to 3.8.

Modify the version scheme used for our cekit maven modules.

The prior scheme, dating from the days when we maintained RHEL7
and RHEL8 images in the same branch, encoded a triplet of RHEL
major version; Maven version; JDK major version into a single
version string, conforming to a specification that limited us
to integers separated by '.'.

E.g.
    8.6.3.6.11 => RHEL 8.6, Maven 3.6, JDK11

Now that we are separating out image sources by RHEL major version,
we can get rid of the first part of the version string to simplify
things.

Since this commit replaces a single maven version across the whole
branch (3.6) with another (3.8), we also do not need to mention it
in the module version. Instead, just mention the JDK major version.

singleton-jdk module no longer needed for the ubi9/openjdk-17 image.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>