[OPENJDK-1510] accept integers for JAVA_*_MEM_RATIO
Ensure that JAVA_INITIAL_MEM_RATIO and JAVA_MAX_MEM_RATIO can accept
Integer-format numbers, as they could prior to the move from calculating
-Xmx/ms to -XX:Max/InitialRAMPercentage.

For JDK8, -XX:Max/InitialRAMPercentage require their parameter to be
formatted as a float; i.e., with a trailing decimal. JDK11+ will accept
a number without the decimal.

To ensure backwards compatibility for users who specified the
environment variables in an integer-format, process the supplied value
with bash's built-in "printf" and output a float.

Since we are now processing the value, rather than passing to the JVM
unaltered, we need to decide upon the precision that we allow. Otherwise
we risk truncating the user-supplied value without warning. Specify that
only the whole-number portion of the value is used, and achieve that
with the appropriate printf format specifier.

We documented that JAVA_MAX_MEM_RATIO=0 disables calculating a JVM flag.
Add a test for this, as well as the other behaviour described above.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>