[OPENJDK-1548] test for cleaned Maven environment
Perform an S2I 'build' of a test Maven project. Define
MAVEN_ARGS for two purposes:

1. so that it exists in the environment prior to invoking Mvn
2. In order to specify the phase 'validate'

The preferred way to achieve (2) is by defining MAVEN_S2I_GOALS,
but defining MAVEN_ARGS for (1) overrides that value.

Please note the chicken-and-egg problem with defining this test. The
test needs to clone the Maven project from a git repository. Until the
PR introducing the test is merged, it does not exist in the main URI.
Therefore I have referenced my GitHub fork from which the PR is being
raised. If/once accepted, a future patch could change 'jmtd' to
'jboss-container-images' and 'OPENJDK-1549-maven-args-take-2' to 'ubi9'
so that the test tracks any changes made to the project in the main
line.

A future enhancement to behave-test-steps may be to permit S2I builds
from a local path relative to the image source repository rather than
requiring an absolute URI to clone from. That would remove the need for
the above.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>