[OPENJDK-1549] clear MAVEN_ARGS in Maven's environment
Future versions of Maven will honour the environment variable MAVEN_ARGS
(using it in much the same way we do). It's possible that this would
result in two copies of arguments being supplied to Maven. To prevent
that, unset MAVEN_ARGS in the environment in which Maven runs, after
expanding it as command-line options.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>