[OPENJDK-2009] Move JAVA_OPTIONS specification to s2i/bash module
JAVA_OPTIONS is an old deprecated variable that was replaced by
JAVA_OPTS. If JAVA_OPTS is unset, but JAVA_OPTIONS set, it should
be used to populate JAVA_OPTS, for the images that support it.

The definition of JAVA_OPTIONS was in modules/jvm/api, but the
implementation is in modules/s2i/bash. The former module is included
by the runtime images, that don't use the latter; thus, at the moment
documentation for the runtime images includes JAVA_OPTIONS but they
have never supported it.

Move the definition of JAVA_OPTIONS into the module that implements
it. This removes the definition/documentation from the runtime images.

Rework the descriptions.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>