[OPENJDK-2145] Update/add usage: labels to UBI8 images
Add a usage: label to JDK11-runtime and JDK8-runtime, pointing at
<https://jboss-container-images.github.io/openjdk/>, which is
probably the best option for the time being. We don't have specific
Red Hat customer documentation for runtime images.

Alter the usage: label for JDK17 to
<https://jboss-container-images.github.io/openjdk/>, as the old
value was for OpenShift 3.x era documentation.

Adjust the usage: label for JDK11 to a fixed URI for the same
document (It changed at some point in the past and we didn't
notice sooner.)

For all UBI8 images, additionally define a second label
org.opencontainers.image.documentation with the same value, following
the OpenContainers Spec, Pre-Defined Annotation Keys:
<https://specs.opencontainers.org/image-spec/annotations/>

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>