User module: parameterize UID, USER, GECOS
Define environment variables capturing the values for the user's name
(USER), uid (UID) and user-name/comment/gecos (GECOS, see passwd(5))
in the jboss.container.user module.

Replace hard-coded references to those values in scripts across all the
other modules. For those modules which now use these variables, add an
explicit dependency on jboss.container.user if it wasn't already
present.

The primary reason for this change is to reduce the delta between the
ubi8 and ubi9 branches, which will avoid merge conflicts for future
backports from ubi9 to ubi8. The secondary reason is to reduce the
risk of a discrepancy for the values across the scripts.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>