Update behave test label checks to check values
Don't test for 'release' or 'architecture' labels which we don't set
in our image sources anyway.

Check for strings in the values of the 'version' and 'name' labels,
to be sure we've set them and not just inherited them.

Split the label-check test into those which apply to all images,
and those which are builder-only.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>