Serially copy JARs when flattening libdir
For some applications, multiple copies of JARs may exist in
JAVA_LIB_DIR. If >1 of those filenames end up collated into the same
`cp` invocation, the result will fail ("cp: will not overwrite
just-created...)

The solution is to serially invoke cp, one per library. This will
cause duplicate filenames to overwrite, which is fine if they are
the same library, but might mask a different problem if they aren't.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>