[OPENJDK-1165] install tar in the runtime images
`oc rsync` requires either `tar` or `rsync` to be present in the
container images. `oc cp` requires `tar`. Install `tar` in the
runtime images so that both `oc` commands function. This results
in a negligible (+0.6MiB) increase in container size.

    <https://issues.redhat.com/browse/OPENJDK-1165>

`tar` and `rsync` are already present in the builder images.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>