[OPENJDK-2587] Ensure pkg-update runs before tzdata module
The tzdata module, which calls "microdnf reinstall tzdata", may
fail if the base image has an installed tzdata RPM that is no
longer available on the mirrors / is out-of-date. For example

    base image installed RPM: tzdata-2023d-1.el8.noarch
    attempt to build ubi8-openjdk-17 (where tzdata module is first)
    "Installed package tzdata-2023d-1.el8.noarch not available"

The fix is to move the pkg-update module to be earlier. Do this
for all images, for consistency, place pkg-update first. (This
matches the ubi9 images)

I have not included a unit test as I am not sure how to approach
mocking older/newer RPM versions in a controllable way.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>