[OPENJDK-2587] tzdata module depend upon pkg-update
Document (as a comment) the issue of ordering pkg-update and
tzdata modules. Depend upon the pkg-update module from tzdata,
to ensure it runs first always. This has no effect on the
container images as they stand now, since they have pkg-config
occurring first already, but if that changed, this would catch
it.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>